+++ src/components/class.jl
@@ -11,18 +11,20 @@ end
 
 function define_class_component(
     mod::Module,
-    class::Symbol,
-    Class::Symbol,
+    singular::Symbol,
+    Singular::Symbol,
+    plural::Symbol,
+    Plural::Symbol,
     short_prefix::Symbol,
 )
-    Class_ = Symbol(Class, :_) # Blueprints module name.
-    _Class = Symbol(:_, Class) # Component type name.
-    sclass, sClass, sp = Meta.quot.((class, Class, short_prefix))
+    Plural_ = Symbol(Plural, :_) # Blueprints module name.
+    _Plural = Symbol(:_, Plural) # Component type name.
+    s, S, sp = Meta.quot.((plural, Plural, short_prefix))
     xp = quote
 
         # ==================================================================================
         # Blueprints for the component.
-        module $Class_
+        module $Plural_
         import EcologicalNetworksDynamics:
             Blueprint, Framework, Networks, GraphDataInputs, @blueprint
         const F = Framework
@@ -46,7 +48,7 @@ function define_class_component(
             Names(names::Vector{Symbol}) = new(names)
         end
 
-        @blueprint Names "raw $($sclass) names"
+        @blueprint Names "raw $($s) names"
         export Names
 
         # Forbid duplicates (triangular check).
@@ -56,14 +58,14 @@ function define_class_component(
                 for j in (i+1):length(names)
                     b = names[j]
                     a == b &&
-                        G.checkfails("$($sClass) $i and $j are both named $(repr(a)).")
+                        G.checkfails("$($S) $i and $j are both named $(repr(a)).")
                 end
             end
         end
 
         # Expand into a new compartment.
         F.expand!(raw, bp::Names) = expand!(raw, bp.names)
-        expand!(raw, names) = Networks.add_class!(raw, $sclass, names)
+        expand!(raw, names) = Networks.add_class!(raw, $s, names)
 
         #-----------------------------------------------------------------------------------
         # Construct from a plain number and generate dummy names.
@@ -71,7 +73,7 @@ function define_class_component(
         mutable struct Number <: Blueprint
             n::UInt
         end
-        @blueprint Number "number of $($sclass)"
+        @blueprint Number "number of $($s)"
         export Number
 
         F.expand!(raw, bp::Number) = expand!(raw, [Symbol($sp, i) for i in 1:bp.n])
@@ -81,24 +83,24 @@ function define_class_component(
         # ==================================================================================
         # The component itself and generic blueprints constructors.
 
-        @component $Class{Internal} blueprints($Class_)
+        @component $Plural{Internal} blueprints($Plural_)
 
         # Build from a number or default to names.
-        (::$_Class)(n::Integer) = $Class.Number(n)
-        (::$_Class)(names) = $Class.Names(names)
+        (::$_Plural)(n::Integer) = $Plural.Number(n)
+        (::$_Plural)(names) = $Plural.Names(names)
 
         # ==================================================================================
         # Display.
 
-        function Framework.shortline(io::IO, model::Model, ::$_Class)
-            names = model.$class._names
+        function Framework.shortline(io::IO, model::Model, ::$_Plural)
+            names = model.$plural._names
             n = length(names)
-            print(io, "$($sClass): $n ($(join_elided(names, ", ")))")
+            print(io, "$($S): $n ($(join_elided(names, ", ")))")
         end
 
     end
     mod.eval.(xp.args)
-    define_class_properties(mod, class, Class, :(depends($Class)))
+    define_class_properties(mod, plural, Plural, singular, Singular, :(depends($Plural)))
 end
 
 # ==========================================================================================
@@ -117,16 +119,18 @@ end
 
 function define_class_properties(
     mod::Module,
-    class::Symbol,
-    Class::Symbol,
+    singular::Symbol,
+    Singular::Symbol,
+    plural::Symbol,
+    Plural::Symbol,
     deps::Expr, # As in a regular call to @method.
 )
-    sclass = Meta.quot(class)
-    M = Symbol(Class, :Methods) # Create submodule to not pollute invocation scope..
+    s = Meta.quot(plural)
+    M = Symbol(Plural, :Methods) # Create submodule to not pollute invocation scope..
     m = :(mod($mod)) # .. but still evaluate dependencies within the invocation module.
     xp = quote
 
-        @propspace $class
+        @propspace $plural
 
         module $M
         import EcologicalNetworksDynamics: Internal, Model, Networks, Framework, Views
@@ -136,23 +140,25 @@ function define_class_properties(
 
         # Nodes counts and nodes labels.
         # The 'ref' variant is more efficient but unexposed.
-        get_number(m::Internal) = n_nodes(m, $sclass)
-        ref_names(m::Internal) = class(m, $sclass).index.reverse
-        get_names(::Internal, m::Model) = nodes_names_view(m, $sclass)
-        @method $m $M.get_number $deps read_as($class.number)
-        @method $m $M.ref_names $deps read_as($class._names)
-        @method $m $M.get_names $deps read_as($class.names)
+        get_number(m::Internal) = n_nodes(m, $s)
+        ref_names(m::Internal) = class(m, $s).index.reverse
+        get_names(::Internal, m::Model) = nodes_names_view(m, $s)
+        @method $m $M.get_number $deps read_as($plural.number)
+        @method $m $M.ref_names $deps read_as($plural._names)
+        @method $m $M.get_names $deps read_as($plural.names)
 
         # Ordered index.
-        ref_index(m::Internal) = class(m, $sclass).index.forward
+        ref_index(m::Internal) = class(m, $s).index.forward
         get_index(m::Internal) = deepcopy(ref_index(m))
-        @method $m $M.ref_index $deps read_as($class._index)
-        @method $m $M.get_index $deps read_as($class.index)
+        indices(m::Internal) = Networks.node_indices(m, $s)
+        @method $m $M.ref_index $deps read_as($plural._index)
+        @method $m $M.get_index $deps read_as($plural.index)
+        @method $m $M.indices $deps read_as($plural.indices)
 
         # Mask within parent class.
         mask(i::Internal, m::Model) =
-            nodes_mask_view(m, ($sclass, class(i, $sclass).parent))
-        @method $m $M.mask $deps read_as($class.mask)
+            nodes_mask_view(m, ($s, class(i, $s).parent))
+        @method $m $M.mask $deps read_as($plural.mask)
 
         end
     end